requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
idna==3.4
requests==2.31.0
smsir-python==1.0.7
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
